Move node JS utilties
