+++ src/_shared/js/utilities/node.js
