+++ src/popup-view/js/GroupsFrame.js
@@ -1,6 +1,6 @@
 import { Frame } from "./Frame.js";
 import GroupDetailFrame from "./GroupDetailFrame.js";
-import { getElementNodeFromString } from "../../_shared/js/models/Node.js";
+import { getElementNodeFromString } from "../../_shared/js/utilities/node.js";
 import { getPluralForm } from "../../js/_share/utils.js";
 
 class GroupsFrame extends Frame {
